[jacob] first crack at ray serve - not working
