+++ backend/Dockerfile
@@ -1,5 +1,7 @@
-FROM python:3.8
+# FROM python:3.8
+FROM rayproject/ray:latest
 
 WORKDIR /app
 ADD requirements.txt /tmp
-RUN pip install -r /tmp/requirements.txt && rm /tmp/requirements.txt
+RUN pip install -r /tmp/requirements.txt 
+# && rm /tmp/requirements.txt
