+++ backend/app/main.py
@@ -1,73 +1,69 @@
-import os
-from neo4j import GraphDatabase
-import json
-import uvicorn
+import requests
+import ray
+from ray import serve
 from fastapi import FastAPI
 
-from uuid import UUID
-
-import sqlalchemy as sa
-from fastapi import Depends, Header, HTTPException
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import Session
-from starlette.status import HTTP_403_FORBIDDEN, HTTP_404_NOT_FOUND
-
-from fastapi_utils.api_model import APIMessage, APIModel
-from fastapi_utils.cbv import cbv
-from fastapi_utils.guid_type import GUID
-from fastapi_utils.inferring_router import InferringRouter
-
-# CORS stuff
-from fastapi.middleware.cors import CORSMiddleware
-
-from .queries import *
-from .models import Subgraph
+# ignore_reinit_error
 
 app = FastAPI()
 
-app.add_middleware(
-    CORSMiddleware,
-    allow_origins=["*"],
-    allow_credentials=True,
-    allow_methods=["*"],
-    allow_headers=["*"],
-)
+ray.init(ignore_reinit_error=True, include_dashboard=False)# 
+# ray.init(address='ray://ray-head:10001', )#address="auto", namespace="summarizer")
+# 'host': None
+serve.start(http_options={'host': None, 'port': 8009})
 
-router = InferringRouter()
+serve_handle = None
 
-# @app.on_startup
+# @app.on_event('startup')
+# async def startup():
+#     # ray.init(ignore_reinit_error=True)# 
+#     # # ray.init(address='ray://ray-head:10001', )#address="auto", namespace="summarizer")
+#     # serve.start(http_options={'host': None})# host=None, port=8001)#http_options={"http_host": None})# detached=True)
+#     Counter.deploy()# _blocking=True)
 
-@cbv(router)
-class App:
-    def __init__(self, uri: str = None):
-        uri = os.getenv('NEO4J_URI', "neo4j://localhost:7687")
-        auth = tuple(os.getenv('NEO4J_AUTH', 'neo4j/ReadThread').split('/'))
-        self.driver = GraphDatabase.driver(uri, auth=auth)
-        self.session = self.driver.session()
-        # self.session.write_transaction(runFullTextIdx)
+#     # global serve_handle
+#     # serve_handle = Counter.get_handle(sync=False)
 
-    @router.get('/friends/{person}')
-    def get_friends(self, person: str):
-        friends = self.session.read_transaction(get_friends_of, person)
-        return friends
+@app.on_event("shutdown")  # Code to be run when the server shuts down.
+async def shutdown_event():
+    serve.shutdown()  # Shut down Ray Serve.
 
-    @router.get('/allpeople')
-    def get_all_people(self):
-        people = self.session.read_transaction(get_all)
-        return people
 
-    @router.get('/search')
-    def full_text_search(self, q: str = ''):
-        if not q:
-            return []
-        return self.session.read_transaction(text_search, q)
+@serve.deployment(route_prefix="/hello")
+# @serve.deployment
+@serve.ingress(app)
+class Counter:
+    def __init__(self):
+        self.count = 0
 
-    @router.post("/subgraph")
-    async def subgraph(self, seeds: Subgraph):
-        # print(f'{k}-hop subgraph for {seeds}')
-        print(f'Subgraph for {seeds}')
-        graph = self.session.read_transaction(get_subgraph_json, seeds)
-        print(graph)
-        return graph
+    def get(self):
+        return {"count": self.count}
 
-app.include_router(router)
+    @app.get("/incr")
+    def incr(self):
+        self.count += 1
+        return {"count": self.count}
+
+    @app.get("/decr")
+    def decr(self):
+        self.count -= 1
+        return {"count": self.count}
+
+@app.get("/")
+async def get():
+    # print(serve_handle)
+    return await Counter.get_handle().get.remote()
+
+# @app.get("/incr")
+# def incr():
+#     self.count += 1
+#     return {"count": self.count}
+
+# @app.get("/decr")
+# def decr():
+#     self.count -= 1
+#     return {"count": self.count}
+
+# print(requests.get('http://localhost:8000').text) # ray-head:6379
+
+Counter.deploy()
