+++ docker-compose.yml
@@ -35,6 +35,7 @@ services:
     environment:
       - NEO4J_URI=neo4j://neo4j:7687
       - NEO4J_AUTH=${NEO4J_AUTH}
+    # command: python /app/app/main.py # 
     command: uvicorn app.main:app --host 0.0.0.0 --reload
     depends_on:
       - neo4j
@@ -55,20 +56,23 @@ services:
     volumes:
       - ./web-react/src:/app/src
 
-  # api:
-  #   build: ./api
-  #   ports:
-  #     - 4001:4001
-  #   environment:
-  #     - NEO4J_URI=bolt://neo4j:7687
-  #     - NEO4J_USER=neo4j
-  #     - NEO4J_PASSWORD=letmein
-  #     - GRAPHQL_LISTEN_PORT=4001
-  #     - GRAPHQL_URI=http://api:4001/graphql
+  ray-head:
+    image: rayproject/ray:latest
+    shm_size: '2gb'
+    entrypoint: [ 'ray']
+    command: ['start', '--head', '--port', '6379', '--include-dashboard', 'true', '--dashboard-host', '0.0.0.0', '--block']
+    ports:
+      - 6379:6379
+      - 8265:8265
 
-  #   links:
-  #     - neo4j
-  #   depends_on:
-  #     - neo4j
-  #   volumes:
-  #     - ./api:/app
+  # 
+  ray-worker:
+    image: rayproject/ray:latest
+    shm_size: '2gb'
+    entrypoint: [ 'ray']
+    command: ['start', '--node-ip-address', 'ray-worker', '--address', 'ray-head:6379', '--block'] # '--object-manager-port', '12345', '--node-manager-port', '12346', 
+    # ports:
+    #   - 12345:12345
+    #   - 12346:12346
+    depends_on:
+      - "ray-head"
